Update 0.2
On startup, Renoir Mobile Tuning now prompts for for admin privileges (we are modifying PCI config data so it needs it anyways).
Updated default values to match the Asus Zephyrus G14.
Boost duration numbers may not be right but
Changed some of the limits to better match what is realistic to get out of these laptops.
Lowered the minimum power limits down to 5w for the heck of it.
Reduced the max temp limit to 97c as most laptops won’t allow for application of settings above this to actually work.
Fixed a bug caused the current limit maximum being 1535A instead of 135A (oops).
Updated the section on what each power limit does